+++ backend/src/main/java/com/commitmate/re_cord/domain/mypage/controller/APiV1MyPageUserController.java
@@ -1,0 +1,13 @@
+package com.commitmate.re_cord.domain.mypage.controller;
+
+
+import lombok.RequiredArgsConstructor;
+import org.springframework.web.bind.annotation.RequestMapping;
+import org.springframework.web.bind.annotation.RestController;
+
+@RestController
+@RequiredArgsConstructor
+@RequestMapping("/mypage")
+
+public class APiV1MyPageUserController {
+}
