+++ backend/src/main/java/com/commitmate/re_cord/domain/mypage/service/MyPageUserService.java
@@ -1,0 +1,14 @@
+package com.commitmate.re_cord.domain.mypage.service;
+
+
+import lombok.RequiredArgsConstructor;
+import lombok.extern.slf4j.Slf4j;
+import org.springframework.stereotype.Service;
+
+@Service
+@RequiredArgsConstructor
+@Slf4j
+public class MyPageUserService {
+
+    // 추후 유저 정보 추가 예정
+}
